modules/java/main/src/test/java/com/mooltiverse/oss/nyx/services/changelog/ChangelogTest.java
/*
 * Copyright 2020 Mooltiverse
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.mooltiverse.oss.nyx.services.changelog;

import static org.junit.jupiter.api.Assertions.*;

import java.util.Map;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.EnumSource;

import com.mooltiverse.oss.nyx.services.Service;

@DisplayName("Changelog")
public class ChangelogTest {
    @Nested
    @DisplayName("Instance")
    class InstanceTest {
        @Test
        public void exceptionWithNullOptions()
            throws Exception {
            assertThrows(NullPointerException.class, () -> Changelog.instance(null));
        }

        @Test
        public void instanceWithEmptyOptions()
            throws Exception {
            Changelog service = Changelog.instance(Map.<String,String>of());
            assertNotNull(service);
        }
    }

    @Nested
    @DisplayName("Supports")
    class SupportsTest {
        @ParameterizedTest(name = "Changelog.instance().supports(''{0}'') == true")
        @EnumSource(Service.Feature.class)
        public void supportAnyFeature(Service.Feature feature)
            throws Exception {
            if (Service.Feature.ASSET.equals(feature))
                assertTrue(Changelog.instance(Map.<String,String>of()).supports(feature));
            else assertFalse(Changelog.instance(Map.<String,String>of()).supports(feature));
        }
    }

    @Nested
    @DisplayName("Asset Service")
    class AssetServiceTest {
        @Test
        @DisplayName("Changelog.instance().buildAsset()")
        public void buildAsset()
            throws Exception {

            // TODO: implement this test
        }
    }
}
